DEV: rename core file in basis module
